server/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
google-auth==2.1.0
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.0
requests==2.26.0
rsa==4.7.2
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
